debt - remove rename/common
