+++ src/vs/workbench/test/electron-browser/api/extHostLanguageFeatures.test.ts
@@ -34,7 +34,7 @@ import { getOccurrencesAtPosition } from 'vs/editor/contrib/wordHighlighter/comm
 import { provideReferences } from 'vs/editor/contrib/referenceSearch/common/referenceSearch';
 import { getCodeActions } from 'vs/editor/contrib/quickFix/common/quickFix';
 import { getWorkspaceSymbols } from 'vs/workbench/parts/search/common/search';
-import { rename } from 'vs/editor/contrib/rename/common/rename';
+import { rename } from 'vs/editor/contrib/rename/browser/rename';
 import { provideSignatureHelp } from 'vs/editor/contrib/parameterHints/common/parameterHints';
 import { provideSuggestionItems } from 'vs/editor/contrib/suggest/common/suggest';
 import { getDocumentFormattingEdits, getDocumentRangeFormattingEdits, getOnTypeFormattingEdits } from 'vs/editor/contrib/format/common/format';
